Structural refactoring: decompose mode files, improve naming
This refactoring was guided and verified by Regrets fingerprint-based
regression testing. All 7 clusters pass with ZERO behavioral changes.

## Changes

### DECOMPOSITION
- Extract parser logic from general-use.js into general-use-parser.js
- Extract parser logic from classical.js into classical-parser.js
- Extract parser logic from beleriand.js into beleriand-parser.js
- Original files now delegate to parser modules and re-export everything

### NAMING
- parseWord → parseTengwarWord (more specific)
- parseColumn → parseTengwarColumn (Tengwar-specific column)
- parseTengwa → matchConsonant (describes what it does)
- parseFollowingS/parseFollowing → parseFollowingSilme (S = silme)
- parseTehta (classical) → parseVowel (more descriptive)

### COHESION
- Parser functions now live in their own modules
- Mode files contain only: name, makeOptions, transcribe, encode, parse
- Shared utilities (parser.js, punctuation.js, etc.) unchanged

### SINGLE RESPONSIBILITY
- Each parser file has one job: parsing text for its mode
- Each mode file has one job: orchestrating the mode's API

### REDUCE COUPLING
- Parser files import from shared utilities only
- Mode files import their parser module
- No cross-mode imports

## Verification

All 4 verifications GREEN after refactoring:
- V1: Regrets clusters all GREEN (7/7)
- V2: Raw output matches KEBENARAN 1 (pre-refactor ground truth)
- V3: Fingerprints match KEBENARAN 2 (pre-refactor contract)
- V4: Chain hashes match (3/3 chains)
- Drift: PASS+STABLE across 5 runs
- Project tests: PASS

## KEBENARAN 1 vs Final Output
All raw outputs are identical to pre-refactoring state.

## Fingerprints Before/After
- general-use-encode: 2heyimj → 2heyimj (identical)
- general-use-encode-black-speech: 3guncp2 → 3guncp2 (identical)
- general-use-encode-reverse-curls: 642tn7o → 642tn7o (identical)
- general-use-transcribe-plain: 5xgljhq → 5xgljhq (identical)
- classical-encode: 3w6m0l5 → 3w6m0l5 (identical)
- classical-encode-early: s3dh268 → s3dh268 (identical)
- beleriand-encode: 20b9l5h → 20b9l5h (identical)

## Chain Hashes Before/After
- general-use-to-beleriand: 187ot9u → 187ot9u (identical)
- black-speech-reverse-curls: 3c1yvbj → 3c1yvbj (identical)
- classical-vs-early: 2k3nbd0 → 2k3nbd0 (identical)
