+++ regrets/adapters/classical-early.js
@@ -1,0 +1,5 @@
+"use strict";
+var Classical = require("../../classical");
+module.exports = function encodeClassical(text) {
+    return Classical.encode(text, { classicalR: true, classicalH: true });
+};
