+++ regrets/adapters/general-use-black-speech.js
@@ -1,0 +1,5 @@
+"use strict";
+var GeneralUse = require("../../general-use");
+module.exports = function encodeBlackSpeech(text) {
+    return GeneralUse.encode(text, { language: "black-speech" });
+};
