+++ regrets/adapters/general-use-reverse-curls.js
@@ -1,0 +1,5 @@
+"use strict";
+var GeneralUse = require("../../general-use");
+module.exports = function encodeReverseCurls(text) {
+    return GeneralUse.encode(text, { reverseCurls: true });
+};
