+++ regrets/adapters/general-use-transcribe-plain.js
@@ -1,0 +1,5 @@
+"use strict";
+var GeneralUse = require("../../general-use");
+module.exports = function transcribePlain(text) {
+    return GeneralUse.transcribe(text, { plain: true });
+};
